Mostly finalized 'Result' type and library; conversion begun
Much work to be done, and many questions to be answered.

First idea: reducing duplicate binding by splitting binding into "binding" and "solidification", and providing methods like a ".BindSolid()" which would both bind and instantly solidify a bound node, indicating a lack of contextual requirements of the AST node. This divide would allow for context-dependent or re-evaluated nodes to be split into a context-independent part to be computed only once and a context-dependent part to be run a variable number of times.

Another question to consider: should nodes be records or classes? if they were classes, modifying tree structure would be trivial. this would also allow for bound-node-completion to return a new instance of the bound node to be inserted in place of "old" versions before completion.